refactor(webhooks): drop .common. segment from model package
Orkes uses io.orkes.conductor.webhook.model (no .common. segment) even
though the files live in the common module. Match that layout: package
is org.conductoross.conductor.webhook.model.

Files still live in common/src/main/java; only the package path changes.
