+++ core/src/main/java/org/conductoross/conductor/dao/webhook/WebhookDAO.java
@@ -15,8 +15,8 @@ package org.conductoross.conductor.dao.webhook;
 import java.util.List;
 import java.util.Map;
 
-import org.conductoross.conductor.common.webhook.model.IncomingWebhookEvent;
-import org.conductoross.conductor.common.webhook.model.WebhookConfig;
+import org.conductoross.conductor.webhook.model.IncomingWebhookEvent;
+import org.conductoross.conductor.webhook.model.WebhookConfig;
 
 /**
  * Persists webhook configurations, matcher indexes, and inbound event audit records.
