+++ webhooks/src/main/java/org/conductoross/conductor/tasks/webhook/InMemoryWebhookDAO.java
@@ -18,8 +18,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.conductoross.conductor.common.webhook.model.IncomingWebhookEvent;
-import org.conductoross.conductor.common.webhook.model.WebhookConfig;
+import org.conductoross.conductor.webhook.model.IncomingWebhookEvent;
+import org.conductoross.conductor.webhook.model.WebhookConfig;
 import org.conductoross.conductor.dao.webhook.WebhookDAO;
 
 /**
